Moved flipCard functionto utils.
