+++ src/common/index.js
@@ -2,4 +2,5 @@ export * from './components/Card'
 export * from './components/cardFaces.js'
 
 export * from './utils/doubleArray.js'
+export * from './utils/flipCard.js'
 export * from './utils/shuffleArray.js'
