+++ src/common/utils/flipCard.js
@@ -1,0 +1,8 @@
+export const flipCard = card => {
+  const newCard = {
+    name: card.name,
+    face: card.face,
+    isFaceUp: !card.isFaceUp
+  };
+  return newCard;
+};
